test(Router): add test for Router::Mode::route

+++ src/serialist/core/generatives/operator.h
@@ -333,7 +333,7 @@ private:
 
 // ==============================================================================================
 
-template<typename FloatType = float>
+template<typename FloatType = double>
 struct OperatorWrapper {
     using Keys = OperatorNode::Keys;
